Align shared library interfaces and data access
Align parser, collection, sorting, storage, and opaque-file interfaces
with their implementations. Carry typed adapters, const accessors,
ownership corrections, and all affected callers together.

Keep filename conversion from modifying stored strings and preserve the
ownership contracts at PIG and raw-byte boundaries. Requires the
preceding public-header and linkage repairs.

+++ libs/UtilitiesLib/src/components/MemRef.c
@@ -5,7 +5,7 @@
 #define MEMREF_TAG 0xC0DF15CE
 
 typedef struct MemRef {
-    volatile int reference_count;
+    volatile LONG reference_count;
     U32 tag;
 } MemRef;
